requirements.txt
# Core RAG and LLM dependencies
langchain==0.1.0
langchain-community==0.0.13

# Vector database for local storage
chromadb==0.4.22

# Embeddings
sentence-transformers==2.3.1

# Document processing
pypdf==4.0.0
python-docx==1.1.0
openpyxl==3.1.2

# Local LLM support
llama-cpp-python==0.2.27

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
